Call out to akmods-ostree-post from %post
We need this for silverblue, so that we can build the
module in %post which is run during construction of the image.

To work with previous versions of akmods we check for
/usr/sbin/akmods-ostree-post before running it.

kmodtool: Forward %dist if set

This allows one to build .fc30 packages on .fc29. Otherwise everything
of the package will be .fc30 except the .srpm that is embedded, which
will be called .fc29.